% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{Importa dati da file TXT in una cartella specificata.}
\usage{
data_import(folder_path = getwd())
}
\arguments{
\item{folder_path}{Percorso della cartella contenente i file TXT da importare. Se non specificato, viene utilizzata la directory di lavoro corrente.}
}
\value{
Un data frame che contiene tutti i dati importati da tutti i file TXT nella cartella specificata.
}
\description{
Questa funzione importa dati da file TXT presenti in una cartella specificata. Ogni file TXT
rappresenta una serie di dati. Il nome del file viene utilizzato come nome del campione.
}
\examples{
# Importa dati dalla cartella "dati"
data <- data_import("dati")

}
